Links on the collection details page go to the collection details form
