+++ app/components/collections/links_component.html.erb
@@ -1,6 +1,8 @@
 <table class="table table-sm mb-5 caption-header">
   <caption>Links to related information
-    <%= render Collections::EditLinkComponent.new(collection_version: collection_version, anchor: 'links', label: 'Edit links to related information') %>
+    <%= link_to edit_collection_version_path(collection_version, anchor: 'links'), aria: { label: 'Edit links to related information' } do %>
+      <span class="fas fa-pencil-alt edit"></span>
+    <% end %>
   </caption>
   <tbody>
     <tr>
