+++ app/views/collection_versions/_collection_version.html.erb
@@ -1,5 +1,6 @@
+<% # The collection details page %>
 <section id="<%= dom_id(collection_version.collection) %>">
-  <%= render Collections::HeaderComponent.new(collection_version: collection_version) %>
+  <%= render Collections::Show::DetailsHeaderComponent.new(collection_version: collection_version) %>
   <%= render Collections::InformationComponent.new(collection_version: collection_version) %>
   <%= render Collections::DetailComponent.new(collection_version: collection_version) %>
   <%= render Collections::LinksComponent.new(collection_version: collection_version) %>
