+++ app/views/collections/_collection.html.erb
@@ -1,5 +1,6 @@
+<% # The collection settings page %>
 <section id="<%= dom_id(collection) %>">
-  <%= render Collections::HeaderComponent.new(collection_version: collection.head) %>
+  <%= render Collections::Show::SettingsHeaderComponent.new(collection_version: collection.head) %>
   <%= render Collections::ReleaseComponent.new(collection: collection) %>
   <%= render Collections::TermsOfUseComponent.new(collection: collection) %>
   <%= render Collections::ParticipantsComponent.new(collection: collection) %>
