+++ app/views/works/index.html.erb
@@ -8,7 +8,7 @@
   <div class="container px-4 px-md-3" id="collection">
     <%= render Collections::Show::TabsComponent.new(collection: @collection) %>
     <section data-controller="work-type">
-      <%= render Collections::HeaderComponent.new(collection_version: @collection.head) %>
+      <%= render Collections::Show::SettingsHeaderComponent.new(collection_version: @collection.head) %>
       <%= render Collections::WorksComponent.new(collection: @collection) %>
     </section>
   </div>
