+++ spec/components/collections/show/details_header_component_spec.rb
@@ -3,7 +3,7 @@
 
 require 'rails_helper'
 
-RSpec.describe Collections::HeaderComponent, type: :component do
+RSpec.describe Collections::Show::DetailsHeaderComponent, type: :component do
   let(:rendered) { render_inline(described_class.new(collection_version: collection_version)) }
 
   context 'with a new, first draft collection' do
@@ -20,6 +20,7 @@ RSpec.describe Collections::HeaderComponent, type: :component do
     it 'does not render the spinner' do
       expect(rendered.to_html).to include 'Depositing'
       expect(rendered.to_html).to include 'fas fa-spinner fa-pulse'
+      expect(rendered.css('a').first['href']).to eq "/collection_versions/#{collection_version.id}/edit"
     end
   end
 end
